Rename PageLocale* events to PageTranslation* events
